Ajustes teste de integração
